requirements.txt
# Core dependencies
Pillow>=10.1.0  # Image processing
PyPDF2>=3.0.0   # PDF manipulation
customtkinter>=5.2.1  # Modern UI widgets
pandas>=2.1.4    # Excel file handling
openpyxl>=3.1.2  # Excel file support
unidecode>=1.3.7  # Text normalization
# Dependencias principales
Pillow>=10.3.0  # Procesamiento de imágenes
customtkinter>=5.2.2  # Interfaz gráfica moderna
pathlib>=1.0.1  # Manejo de rutas
pandas>=2.1.4  # Procesamiento de datos
openpyxl>=3.1.2  # Lectura/escritura de Excel
PyPDF2>=3.0.1  # Manipulación de PDFs

# Dependencias de desarrollo
pytest>=8.3.4  # Testing
coverage>=7.3.0  # Cobertura de código
pylint>=3.3.3  # Análisis estático
black>=23.12.1  # Formateo de código
pyinstaller>=6.3.0  # Generación de ejecutables

# Dependencias opcionales
python-dotenv>=1.0.0  # Variables de entorno
tqdm>=4.66.1  # Barras de progreso
